Debug version of session key plugin for surfacing errors
Debug version of the session key plugin which reverts with descriptive
string errors on permission failures. This can be used in a state
override during simulation, replacing the deployed session key plugin's
code with this version.

To optimize for this use case, all external `SessionKeyPlugin` functions
have been removed other than `executeWithSessionKey` and
`userOpValidationFunction`, which reduces the bytecode size from 15kb to
5kb.